+++ src/components/manage/Widgets/TextareaWidget.stories.jsx
@@ -7,6 +7,16 @@ export const Textarea = WidgetStory.bind({
   widget: TextareaWidget,
 });
 
+export const DefaultValue = WidgetStory.bind({
+  props: {
+    id: 'textarea',
+    title: 'Text area',
+    block: 'block',
+    default: 'Hello world',
+  },
+  widget: TextareaWidget,
+});
+
 export default {
   title: 'Edit Widgets/Textarea',
   component: TextareaWidget,