+++ CHANGELOG.md
@@ -5,16 +5,18 @@
 ### Breaking
 
 - The `ContentsBreadcrumbs` component now renders the whole language name of the language root folder (if any) instead of just the `id` (before: `de`, now: `Deutsch`) @sneridagh
+- The `ObjectWidget` now uses `default` instead of `defaultValue` for fields coming from schema @tiberiuichim
 
 ### Feature
 
 - Improvement of the `ContentsBreadcrumbs` component, add child `ContentsBreadcrumbsRootItem` and `ContentsBreadcrumbsHomeItem` for easy customization of these single elements in projects @sneridagh
 - Add german translation for group membership panel. @ksuess
 - Fix general german translations: Address user polite. Correct 'listing template' to 'listing variant'. Add missing translations. @ksuess
+- Add support for `default` prop to textarea fields @tiberiuichim
 
 ### Bugfix
 
-- Fix ObjectWidget handling of `default` values coming from schemas.
+- Fix ObjectWidget handling of `default` values coming from schemas. @tiberiuichim
 
 ### Internal